Checkbox component clean up
