+++ frontend/src/metabase/core/components/CheckBox/utils.ts
@@ -1,0 +1,3 @@
+export function isEllipsisActive(node: HTMLElement): boolean {
+  return node.offsetWidth < node.scrollWidth;
+}
